requirements.txt
discord.py==2.4.0
redis==5.0.1
audioop-lts

